+++ dnacentersdk/models/validators/v2_3_5_3/jsd_e11daa984f535a08bc1eb01bc84bc399.py
@@ -82,7 +82,7 @@ class JSONSchemaValidatorE11Daa984F535A08Bc1EB01Bc84Bc399(object):
                 },
                 "type": "object"
                 },
-                "interfaceName": {
+                "ipInterfaceName": {
                 "type": "string"
                 },
                 "rfProfile": {
@@ -111,7 +111,7 @@ class JSONSchemaValidatorE11Daa984F535A08Bc1EB01Bc84Bc399(object):
                 ],
                 "type": "string"
                 },
-                "vlanID": {
+                "vlanId": {
                 "type": "string"
                 }
                 },